+++ HISTORY.md
@@ -3,6 +3,7 @@ unreleased
 
   * Add `application/x-keepass2` with extension `.kdbx`
   * Add extension `.mxmf` to `audio/mobile-xmf`
+  * Add extensions from IANA for `application/*+xml` types
   * Add new upstream MIME types
 
 1.42.0 / 2019-09-25